# storage/api-docs/v2-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brillio API",
        "description": "API Backend for Brillio Orientation System",
        "contact": {
            "email": "contact@brillio.africa"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Brillio API Server"
        }
    ],
    "paths": {
...
        "/api/v2/feedback": {
            "post": {
                "tags": [
                    "UserProfiling"
                ],
                "summary": "Enregistre le feedback d'un utilisateur",
                "operationId": "60254be8391ef8b2359366f6467231fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "comment"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "example": "Super service !"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "needs_situation": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/feedback/skip": {
            "post": {
                "tags": [
                    "UserProfiling"
                ],
                "summary": "Reporte le nudge de feedback",
                "operationId": "de6f3380b6a4d6d20247f6756a80ee6f",
                "responses": {
                    "200": {
                        "description": "Nudge reporté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/situation": {
            "post": {
                "tags": [
                    "UserProfiling"
                ],
                "summary": "Enregistre la situation détaillée d'un jeune",
                "operationId": "b308ca4cec1c584f8c5698cb1c7ac8bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "city": {
                                                "type": "string",
                                                "example": "Cotonou"
                                            },
                                            "institution": {
                                                "type": "string",
                                                "example": "Université"
                                            },
                                            "specialization": {
                                                "type": "string",
                                                "example": "Informatique"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Situation mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/situation/skip": {
            "post": {
                "tags": [
                    "UserProfiling"
                ],
                "summary": "Reporte le nudge de situation",
                "operationId": "4b6a66e26172d5f3fcd851b26ba1ca24",
                "responses": {
                    "200": {
                        "description": "Nudge reporté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}